duplicate Z class in each module
