+++ tests/utils/test_json.py
@@ -28,7 +28,25 @@ import pytest
 
 from airflow.datasets import Dataset
 from airflow.utils import json as utils_json
-from tests.serialization import Z
+
+
+class Z:
+    __version__: ClassVar[int] = 1
+
+    def __init__(self, x):
+        self.x = x
+
+    def serialize(self) -> dict:
+        return dict({"x": self.x})
+
+    @staticmethod
+    def deserialize(data: dict, version: int):
+        if version != 1:
+            raise TypeError("version != 1")
+        return Z(data["x"])
+
+    def __eq__(self, other):
+        return self.x == other.x
 
 
 @dataclass
